+++ pyccel/ast/basic.py
@@ -560,6 +560,17 @@ class TypedAstNode(PyccelAstNode):
         """
         return self._order # pylint: disable=no-member
 
+    @property
+    def class_type(self):
+        """
+        The type of the object.
+
+        The Python type of the object. In the case of scalars this is equivalent to
+        the datatype. For objects in (homogeneous) containers (e.g. list/ndarray/tuple),
+        this is the type of the container.
+        """
+        return self._class_type # pylint: disable=no-member
+
     @classmethod
     def static_rank(cls):
         """
@@ -618,6 +629,20 @@ class TypedAstNode(PyccelAstNode):
         """
         return cls._order # pylint: disable=no-member
 
+    @classmethod
+    def static_class_type(cls):
+        """
+        The type of the object.
+
+        The Python type of the object. In the case of scalars this is equivalent to
+        the datatype. For objects in (homogeneous) containers (e.g. list/ndarray/tuple),
+        this is the type of the container.
+
+        This function is static and will return an AttributeError if the
+        class does not have a predetermined order.
+        """
+        return cls._class_type # pylint: disable=no-member
+
     def copy_attributes(self, x):
         """
         Copy the attributes describing a TypedAstNode into this node.
@@ -631,11 +656,12 @@ class TypedAstNode(PyccelAstNode):
         x : TypedAstNode
             The node from which the attributes should be copied.
         """
-        self._shape     = x.shape
-        self._rank      = x.rank
-        self._dtype     = x.dtype
-        self._precision = x.precision
-        self._order     = x.order
+        self._shape      = x.shape
+        self._rank       = x.rank
+        self._dtype      = x.dtype
+        self._precision  = x.precision
+        self._order      = x.order
+        self._class_type = x.class_type
 
 
 #------------------------------------------------------------------------------